串接API - 使用 PUT 資料修改

+++ src/views/CategoryCRUD.vue
@@ -8,6 +8,7 @@ const category = ref({
     categoryid:0,
     categoryname:''
 })
+const btnText = ref('新增')
 
     // const API_URL = 'http://127.0.0.1:8000/api/category/'
     //串接API - GET 資料讀取
@@ -21,7 +22,15 @@ const category = ref({
 
     //串接API - POST 資料新增
     const addCategory = async()=>{
-        // console.log(category.value)
+        //有可能做新增
+        //有可能做修改
+
+        if(category.value.categoryid > 0){
+            //console.log('修改')
+            updateCategory()
+        }else{
+            //console.log('新增')
+            // // console.log(category.value)
         const response = await fetch(API_URL,{
             method:'POST',
             body: JSON.stringify(category.value),
@@ -33,20 +42,48 @@ const category = ref({
         }else{
             alert('新增失敗')
         }
+        }
+
+
+        
     }
 
     //串接 API - 資料刪除
-    const removeCategory = async(category) => {
-      //console.log(category) category/14/
-      const response = await fetch(`${API_URL}${category.categoryid}/`,{
-        method: 'DELETE'
+    const removeCategory = async(item) => {
+
+      if(window.confirm('真的要刪除嗎?')){
+        //console.log(item) category/14/
+        const response = await fetch(`${API_URL}${item.categoryid}/`,{
+                method: 'DELETE'
+            })
+            if(response.ok){
+                loadCategories()
+            }
+      }
+
+     
+    }
+
+    //串接 API - 資料修改
+    const updateCategory = async()=>{
+      const response = await fetch(`${API_URL}${category.value.categoryid}/`,{
+        method: 'PUT',
+        body: JSON.stringify(category.value),
+        headers:{'Content-Type': 'application/json'}
       })
       if(response.ok){
+        clearCategory()
         loadCategories()
       }
     }
 
+    const editCategory = item => {
+       category.value = item
+       btnText.value = "修改"
+    }
+
     const clearCategory = ()=>{
+        btnText.value = "新增"
         category.value = {
             categoryid:0,
             categoryname:''
@@ -64,7 +101,7 @@ const category = ref({
   <li v-for="category in categories" :key="category.categoryid" class="list-group-item d-flex justify-content-between align-items-center">
    {{ category.categoryname }}
    <div>
-    <span class="badge text-bg-primary rounded-pill m-1"><i class="bi bi-pencil-square"></i></span>
+    <span @click="editCategory(category)" class="badge text-bg-primary rounded-pill m-1"><i class="bi bi-pencil-square"></i></span>
     <span @click="removeCategory(category)" class="badge text-bg-danger rounded-pill m-1"><i class="bi bi-trash-fill"></i></span>
    </div>
   </li>
@@ -75,7 +112,7 @@ const category = ref({
     <div class="col-4">
         <div class="input-group mb-3">
   <input type="text" v-model="category.categoryname" class="form-control" placeholder="請輸入景點分類">
-  <button @click="addCategory" class="btn btn-outline-secondary" type="button" id="button-addon2">送出</button>
+  <button @click="addCategory" class="btn btn-outline-secondary" type="button" id="button-addon2">{{btnText}}</button>
 </div>        
     </div>
     <div class="col-4"></div>
